Move StrainPeaks to API
