+++ app/rulesets/osu/performance/pp211112/difficulty.go
@@ -17,21 +17,6 @@ const (
 	CurrentVersion    int     = 20211112
 )
 
-// StrainPeaks contains peaks of Aim, Speed and Flashlight skills, as well as peaks passed through star rating formula
-type StrainPeaks struct {
-	// Aim peaks
-	Aim []float64
-
-	// Speed peaks
-	Speed []float64
-
-	// Flashlight peaks
-	Flashlight []float64
-
-	// Total contains aim, speed and flashlight peaks passed through star rating formula
-	Total []float64
-}
-
 // getStarsFromRawValues converts raw skill values to Attributes
 func getStarsFromRawValues(rawAim, rawAimNoSliders, rawSpeed, rawFlashlight float64, diff *difficulty.Difficulty, attr api.Attributes, _ bool) api.Attributes {
 	aimRating := math.Sqrt(rawAim) * StarScalingFactor
@@ -184,7 +169,7 @@ func CalculateStep(objects []objects.IHitObject, diff *difficulty.Difficulty, ex
 	return stars
 }
 
-func CalculateStrainPeaks(objects []objects.IHitObject, diff *difficulty.Difficulty, experimental bool) StrainPeaks {
+func CalculateStrainPeaks(objects []objects.IHitObject, diff *difficulty.Difficulty, experimental bool) api.StrainPeaks {
 	diffObjects := preprocessing.CreateDifficultyObjects(objects, diff, experimental)
 
 	aimSkill := skills2.NewAimSkill(diff, true, experimental)
@@ -197,7 +182,7 @@ func CalculateStrainPeaks(objects []objects.IHitObject, diff *difficulty.Difficu
 		flashlightSkill.Process(o)
 	}
 
-	peaks := StrainPeaks{
+	peaks := api.StrainPeaks{
 		Aim:        aimSkill.GetCurrentStrainPeaks(),
 		Speed:      speedSkill.GetCurrentStrainPeaks(),
 		Flashlight: flashlightSkill.GetCurrentStrainPeaks(),
